pinctrl/rockchip: separate struct rockchip_pin_bank to a head file
[ Upstream commit e1450694e94657458395af886d2467d6ac3355af ]

Separate struct rockchip_pin_bank to pinctrl-rockchip.h file, which will
be used by gpio-rockchip driver in the future.

Link: https://lore.kernel.org/r/20210816011948.1118959-3-jay.xu@rock-chips.com
Stable-dep-of: 01b4b1d1cec4 ("pinctrl: rockchip: use dedicated pinctrl type for RK3328")
